Удалил класс Main, перенес метод main обратно, удалил TestMain. Добавил тесты
