--- Task_1_1_2/src/test/java/ru/nsu/kozoliy/TestMain.java
@@ -1,10 +1,0 @@
-package ru.nsu.kozoliy;
-
-import org.junit.jupiter.api.Test;
-
-public class TestMain {
-    @Test
-    void testMain () {
-        Main.main(new String[0]);
-    }
-}
